add plan workflow test and remove customize executor test
