+++ tests/covalent_dispatcher_tests/_core/init_test.py
@@ -19,7 +19,52 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 """
-Unit tests for the core functionality of the dispatcher.
+Tests for the core functionality of the dispatcher.
 """
 
-# TODO - Write unit tests
+import cloudpickle as pickle
+import covalent as ct
+
+from covalent_dispatcher._core import _plan_workflow
+from covalent._results_manager import Result
+from covalent._results_manager import results_manager as rm
+
+TEST_RESULTS_DIR = "/tmp/results"
+
+
+def get_mock_result() -> Result:
+    """Construct a mock result object corresponding to a lattice."""
+
+    import sys
+
+    @ct.electron
+    def task(x):
+        print(f"stdout: {x}")
+        print("Error!", file=sys.stderr)
+        return x
+
+
+    @ct.lattice(results_dir=TEST_RESULTS_DIR)
+    def pipeline(x):
+        return task(x)
+
+    pipeline.build_graph(x="absolute")
+    pipeline.check_consumable()
+
+    return Result(
+        lattice=pipeline,
+        results_dir=pipeline.metadata["results_dir"],
+    )
+
+
+def test_plan_workflow():
+    """Test workflow planning method.
+    """
+
+    mock_result = get_mock_result()
+    _plan_workflow(result_object=mock_result)
+
+    # Updated transport graph post planning
+    updated_tg = pickle.loads(mock_result.lattice.transport_graph.serialize(metadata_only=True))
+
+    assert 'results_dir' in updated_tg['nodes'][0]['exec_plan'].execution_args
